Toke should expected id of type Long not Integer

+++ src/main/java/com/meufty/workoutplanner/repository/TokenRepository.java
@@ -12,7 +12,7 @@ public interface TokenRepository extends JpaRepository<Token, Integer> {
             select t from Token t inner join MyUser u on t.user.id = u.id
             where u.id = :userId and (t.expired = false or t.revoked = false )
             """)
-    List<Token> findAllValidTokensByUser(Integer userId);
+    List<Token> findAllValidTokensByUser(Long userId);
 
     Optional<Token> findByToken(String token);
 }
